+++ books_new_life/books/admin.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
-from .models import Books, Categorias, Profile
+from .models import Books, Categorias
 
 
 admin.site.register(Books)
 admin.site.register(Categorias)
-admin.site.register(Profile)
+
 
 # Register your models here.
