+++ books_new_life/miembros/views.py
@@ -3,8 +3,9 @@ from django.shortcuts import render
 from django.views import generic
 from django.contrib.auth.forms import UserCreationForm, PasswordChangeForm
 from django.urls import reverse_lazy
-from .forms import SignUpForm, UserEditForm
+from .forms import SignUpForm, UserEditForm, ProfileForm
 from django.contrib.auth.views import PasswordChangeView
+from django.contrib.auth.models import User
 
 # Create your views here.
 
@@ -32,5 +33,6 @@ class PasswordChangeView(PasswordChangeView):
 
 def ProfileView(request):
     user = request.user
+    form = ProfileForm(instance=user.profile)
 
-    return render(request, 'profile.html', {'user': user})
+    return render(request, 'profile.html', {'user': user, 'form': form})
