+++ books_new_life/templates/profile.html
@@ -40,16 +40,27 @@
 
     <div class="flex">
       <div class="inputBox">
+        {% comment %} {{form.descripcion}} {% endcomment %}
         <span>Nombre de usuario:</span>
         {{user.username}}
         <span>Nombre:</span>
         {{user.first_name|title}}
-      </div>
-      <div class="inputBox">
         <span>Apellido:</span>
         {{user.last_name|title}}
         <span>Correo electronico:</span>
         {{user.email}}
+      </div>
+      <div class="inputBox">
+        <span>Departamento:</span>
+        {{form.departamento|title}}
+        <span>Población:</span>
+        {{form.poblacion|title}}
+        <span>Telefono:</span>
+        {{form.telefono|title}}
+        <span>WhatsApp:</span>
+        {{form.whatsapp|title}}
+        <span>Facebook:</span>
+        {{form.facebook|title}}
       </div>
     </div>
     <a href="" class="delete-btn">Mis libros subidos</a>
